pyproject.toml
[project]
name = "zepp-pc"
version = "0.1.0"
description = "A lightweight PC manager for Amazfit/Zepp smartwatches"
requires-python = ">=3.10"
dependencies = [
    "bleak>=0.22.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pywebview>=5.0.0",
    "cryptography>=43.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
gui = [
    "qtpy>=2.4.0",
    "PyQt6>=6.5.0",
    "PyQt6-WebEngine>=6.5.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
